+++ src/fhapi/system.h
@@ -1,13 +1,13 @@
 /*
 #   clove
 #
-#   Copyright (C) 2016-2020 Muresan Vlad
+#   Copyright (C) 2016-2025 Muresan Vlad
 #
 #   This project is free software; you can redistribute it and/or modify it
 #   under the terms of the MIT license. See LICENSE.md for details.
 */
 #pragma once
 
-#include "../3rdparty/mojoAL/AL/al.h"
+#include "../3rdparty/FH/src/fh.h"
 
-int audio_wav_load(unsigned int buffer, char const* filename, float* outSeconds);
+void fh_system_register(struct fh_program *prog);